bench3

+++ bench3/ddl/ddl.sql
@@ -3,7 +3,7 @@ CREATE TABLE logs (
   device_id    CHAR(15) NOT NULL,
   device_name  VARCHAR(25) NOT NULL,
   device_type  VARCHAR(15) NOT NULL,
-  device_floor TINYINT NOT NULL,
+  device_floor SMALLINT NOT NULL,
   event_type   VARCHAR(15) NOT NULL,
   event_unit   CHAR(1),
   event_value  FLOAT
